DoctorWhoFestival.py
A three-day festival is the official celebration of all things of “Doctor Who”. 
There are three categories of tickets: individual cost is £68, under sixteen is £32.25 and a family cost 42.75. Write a program that asks the user for total numbers in each category, calculate and display the breakdown of each and total prices.