Add files via upload
Write a program that gives information about James Bond films. The program gives you 4 chances to name an actor who has played Bond and then say whether you are right or not. They should use the text below in the messages output (up to the names of the actors and so films given). Finally they should print how well you did giving a score out of 4.